requirements.txt
fastapi
uvicorn 
gunicorn
psycopg2-binary
sqlmodel
mysqlclient
passlib[bcrypt]
pyjwt
python-multipart
itsdangerous
instamojo-wrapper

